feat(google-smart-home): add types for trait and command
