+++ google-smart-home/tests/query.rs
@@ -1,8 +1,8 @@
 mod common;
 
 use google_smart_home::query::request;
+use google_smart_home::query::response;
 use google_smart_home::query::response::Response;
-use google_smart_home::query::response::{self};
 use google_smart_home::Request;
 use google_smart_home::RequestInput;
 use serde_json::json;
